Move ownership of VRDevice components to VRManager.
