+++ src/Engine/Manager/Managers.cpp
@@ -57,5 +57,6 @@ void Hub::ClearKilledComponents() {
         physicsManager->ClearKilledComponents();
         soundManager->ClearKilledComponents();
         scriptManager->ClearKilledComponents();
+        vrManager->ClearKilledComponents();
     }
 }
